split unit and integration tests on build tag
